Multiple input data flow for dummy backend
Fix the multiple input data flow support.
In some cases, the OpenAPI defines both image requests and video
requests, which must be directed to different data flow.
The CL added a MultipleFlowCollect to perform a poll like action
to collect the data from either data flow as long as there is data